+++ WcfBankingService/operation/operations/Deposit.cs
@@ -6,8 +6,8 @@ namespace WcfBankingService.Operation.Operations
     public class Deposit : BankOperation, IBankCommand
     {
         private readonly IAccount _targetAccount;
-        private readonly BigInteger _amount;
-        public Deposit(IAccount targetAccount, BigInteger amount, string operationTitle) : base(operationTitle, amount, "Deposit")
+        private readonly decimal _amount;
+        public Deposit(IAccount targetAccount, decimal amount, string operationTitle) : base(operationTitle, amount, "Deposit")
         {
             _targetAccount = targetAccount;
             _amount = amount;
